rename theme registry
